Move MCPServer data class to a new file for better organization
