--- mcp/src/main/kotlin/predictable/mcp/Main.kt
@@ -1,1 +1,0 @@
-package predictable.mcp
