+++ mcp/src/main/kotlin/predictable/mcp/config/MCPConfig.kt
@@ -7,10 +7,3 @@ data class MCPConfig(
   val servers: Map<String, MCPServer>,
 )
 
-@Serializable
-data class MCPServer(
-  val name: String,
-  val namespace: String,
-  val description: String,
-  val config: ServerConfig
-)
