+++ mcp/src/main/kotlin/predictable/mcp/config/MCPServer.kt
@@ -1,0 +1,9 @@
+package predictable.mcp.config
+
+@Serializable
+data class MCPServer(
+  val name: String,
+  val namespace: String,
+  val description: String,
+  val config: ServerConfig
+)
