--- mcp/src/main/kotlin/predictable/mcp/registry/agent/FindToolsRequest.kt
@@ -1,8 +1,0 @@
-package predictable.mcp.registry.agent
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FindToolsRequest(
-  val query: String,
-)
