--- mcp/src/main/kotlin/predictable/mcp/registry/agent/FindToolsResponse.kt
@@ -1,8 +1,0 @@
-package predictable.mcp.registry.agent
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FindToolsResponse(
-  val tools: List<SelectedTool>,
-)
