--- mcp/src/main/kotlin/predictable/mcp/registry/agent/SelectedTool.kt
@@ -1,9 +1,0 @@
-package predictable.mcp.registry.agent
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SelectedTool(
-  val name: String,
-  val namespace: String,
-)
