--- mcp/src/main/kotlin/predictable/mcp/registry/agent/Toolset.kt
@@ -1,8 +1,0 @@
-package predictable.mcp.registry.agent
-
-import predictable.mcp.config.MCPConfig
-
-sealed class Toolset {
-  data object DockerHub : Toolset()
-  data class Custom(val config: MCPConfig) : Toolset()
-}
