--- mcp/src/main/resources/logback.xml
@@ -1,19 +1,0 @@
-<configuration>
-    <!-- Silence logback initialization status messages -->
-    <statusListener class="ch.qos.logback.core.status.NopStatusListener" />
-
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%d{YYYY-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-    <root level="off">
-        <appender-ref ref="STDOUT"/>
-    </root>
-    <!-- Silence specific loggers -->
-    <logger name="predictable.agent.providers.openai" level="OFF"/>
-    <logger name="io.ktor" level="OFF"/>
-    <logger name="ch.qos.logback" level="OFF"/>
-    <logger name="org.eclipse.jetty" level="OFF"/>
-    <logger name="io.netty" level="OFF"/>
-</configuration>
